Move runtime metric mappings to separate file
